+++ plugins/11_TheWarWithin/localization/enUS.lua
@@ -259,3 +259,7 @@ L['tkaktath_note'] = 'Starts a quest chain to get the {item:224150} mount.'
 L['options_icons_the_unseeming'] = '{achievement:40633}'
 L['options_icons_the_unseeming_desc'] = 'Display location for the {achievement:40633} achievement.'
 L['the_unseeming_note'] = 'Stand in the pool until you have 100x stacks {spell:420847}.'
+
+L['options_icons_you_cant_hang_with_us'] = '{achievement:40634}'
+L['options_icons_you_cant_hang_with_us_desc'] = 'Display locations for the {achievement:40634} achievement.'
+L['you_cant_hang_with_us_note'] = 'Find a {npc:220986} with the {spell:434734} buff, if you stand close to him he will give you the {spell:443190} debuff (1 minute) ({note:Do not attack him!}). Find a {npc:225408} to get 10x stacks of {spell:454666}.'